Askademia/ta-bot/requirements.txt
uagents
google-genai
python-dotenv
pydantic
httpx
pymupdf 
tiktoken 
tqdm
pymongo[srv]
requests
beautifulsoup4
streamlit